Refactoring files (#741)
